+++ app/src/main/java/com/group1/movebetter/repository/Repository.kt
@@ -1,16 +1,53 @@
 package com.group1.movebetter.repository
 
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.google.gson.Gson
 import com.group1.movebetter.model.CityBikes
 import com.group1.movebetter.model.CityBikesNetworkList
 import com.group1.movebetter.model.Departures
 import com.group1.movebetter.model.StaDaStations
 import com.group1.movebetter.model.*
 import com.group1.movebetter.network.RetrofitInstance
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 class Repository {
-    suspend fun getNetworks(): CityBikes
+
+    private val _getResponseNetworks: MutableLiveData<CityBikes> = MutableLiveData()
+    val getResponseNetworks: LiveData<CityBikes>
+        get() = _getResponseNetworks
+
+    private val _getResponseNetwork: MutableLiveData<CityBikesNetworkList> = MutableLiveData()
+    val getResponseNetwork: LiveData<CityBikesNetworkList>
+        get() = _getResponseNetwork
+
+    private val _getResponseArrival: MutableLiveData<Departures> = MutableLiveData()
+    val getResponseArrival: LiveData<Departures>
+        get() = _getResponseArrival
+
+    private val _getResponseStations: MutableLiveData<StaDaStations> = MutableLiveData()
+    val getResponseStations: LiveData<StaDaStations>
+        get() = _getResponseStations
+
+    private val _myResponse: MutableLiveData<EmailBody> = MutableLiveData()
+    val myResponse: LiveData<EmailBody>
+        get() = _myResponse
+
+    private val _myTokens: MutableLiveData<BirdTokens> = MutableLiveData()
+    val myTokens: LiveData<BirdTokens>
+        get() = _myTokens
+
+    private val _myBirds: MutableLiveData<Birds> = MutableLiveData()
+    val myBirds: LiveData<Birds>
+        get() = _myBirds
+
+    suspend fun getNetworks()
     {
-        return RetrofitInstance.apiCityBikes.getNetworks()
+        withContext(Dispatchers.IO){
+            val responseNetworks = RetrofitInstance.apiCityBikes.getNetworks()
+            _getResponseNetworks.value = responseNetworks
+        }
     }
 
     suspend fun getNetworksFiltered(fields: String): CityBikes
@@ -18,33 +55,51 @@ class Repository {
         return RetrofitInstance.apiCityBikes.getNetworksFiltered(fields)
     }
 
-    suspend fun getNetwork(networkId: String): CityBikesNetworkList
+    suspend fun getNetwork(networkId: String)
     {
-        return RetrofitInstance.apiCityBikes.getNetwork(networkId)
+        withContext(Dispatchers.IO){
+            val responseNetwork = RetrofitInstance.apiCityBikes.getNetwork(networkId)
+            _getResponseNetwork.value = responseNetwork
+        }
     }
 
-    suspend fun getStations(): StaDaStations
+    suspend fun getStations()
     {
-        return RetrofitInstance.apiStadaStations.getStations()
+        withContext(Dispatchers.IO){
+            val responseStations = RetrofitInstance.apiStadaStations.getStations()
+            _getResponseStations.value = responseStations
+        }
     }
 
-    suspend fun getArrival(evaId: Long, lookahead: Long): Departures
+    suspend fun getArrival(evaId: Long, lookahead: Long)
     {
-        return RetrofitInstance.apiMarudor.getArrival(evaId, lookahead)
+        withContext(Dispatchers.IO){
+            val getResponseArrival = RetrofitInstance.apiMarudor.getArrival(evaId, lookahead)
+            _getResponseArrival.value = getResponseArrival
+        }
     }
     suspend fun getBirdToken(body: EmailBody) {
         return RetrofitInstance.birdAuthApi.getAuthToken(body)
     }
 
-    suspend fun postMagicToken(body: Token): BirdTokens {
-        return RetrofitInstance.birdAuthApi.postAuthToken(body)
+    suspend fun postMagicToken(body: Token) {
+        withContext(Dispatchers.IO){
+            val response = RetrofitInstance.birdAuthApi.postAuthToken(body)
+            _myTokens.value = response
+        }
     }
 
-    suspend fun refresh(token: String): BirdTokens {
-        return RetrofitInstance.birdAuthApi.refresh(token)
+    suspend fun refresh(token: String) {
+        withContext(Dispatchers.IO){
+            val response = RetrofitInstance.birdAuthApi.refresh(token)
+            _myTokens.value = response
+        }
     }
 
-    suspend fun getBirds(lat: Double, lng: Double, rad: Int, token: String, location: String): Birds {
-        return RetrofitInstance.birdApi.getNearbyBirds(lat, lng, rad, token, location)
+    suspend fun getBirds(lat: Double, lng: Double, rad: Int, token: String, location: String) {
+        withContext(Dispatchers.IO){
+            val response = RetrofitInstance.birdApi.getNearbyBirds(lat, lng, rad, token, location)
+            _myBirds.value = response
+        }
     }
 }
