+++ app/src/main/java/com/group1/movebetter/view_model/controller/CityBikeController.kt
@@ -9,35 +9,18 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
 class CityBikeController (private val viewModelScope: CoroutineScope, private val repository: Repository) {
-    //ehemals MainViewModel:
-
-    private val _getResponseNetworks: MutableLiveData<CityBikes> = MutableLiveData()
-    val getResponseNetworks: LiveData<CityBikes>
-        get() = _getResponseNetworks
-
-    private val _getResponseNetwork: MutableLiveData<CityBikesNetworkList> = MutableLiveData()
-    val getResponseNetwork: LiveData<CityBikesNetworkList>
-        get() = _getResponseNetwork
-
-    private val _getResponseNetworkFiltered: MutableLiveData<CityBikesNetworkList> = MutableLiveData()
 
     fun getNetworks()
     {
         viewModelScope.launch {
-            val responseNetworks = repository.getNetworks()
-            _getResponseNetworks.value = responseNetworks
-
-		/*val responseNetworkFiltered = repository.getNetwork("")
-		getResponseNetworkFiltered.value = responseNetworkFiltered*/
-
+            repository.getNetworks()
         }
     }
 
     fun getNetwork(id: String)
     {
         viewModelScope.launch {
-            val responseNetwork = repository.getNetwork(id)
-            _getResponseNetwork.value = responseNetwork
+            repository.getNetwork(id)
         }
     }
